+++ home/index.html
@@ -41,21 +41,10 @@
 
   <h3>How to use</h3>
   <ul>
-    <li class="bullet"><strong>[command..]</strong>
+    <li class="bullet"><strong>Generate screenshots for a JavaScript sketch</strong>
       <ul>
-        <li>[A description of the commands.....]<br><div class="code">$ [docome command...]<br>=> [the output on terminal]</div></li><br>[what should the user expect to get...]
-      </ul>
-    </li>
-
-    <li class="bullet"><strong>[command..]</strong>
-      <ul>
-        <li>[A description of the commands.....]<br><div class="code">$ [docome command...]<br>=> [the output on terminal]</div></li><br>[what should the user expect to get...]
-      </ul>
-    </li>
-
-    <li class="bullet"><strong>[command..]</strong>
-      <ul>
-        <li>[A description of the commands.....]<br><div class="code">$ [docome command...]<br>=> [the output on terminal]</div></li><br>[what should the user expect to get...]
+        <li>While in the project main folder, run the following command<br><div class="code">$ docode --screenshots</div></li>A new folder with the screenshots will be created at the same location.<br><br>
+        <li>To generate for a project on another folder, run the following command<br><div class="code">$ docode --screenshots=[path_to_project_main_folder]</div></li>A new folder with the screenshots will be created in the project main folder.
       </ul>
     </li>
