--- test.js
@@ -1,7 +1,0 @@
-#!/usr/bin/env node
-
-console.log('hi');
-
-
-
-console.log(__dirname);
